added data-fields to settings

+++ backend/tracker/infrastructure/dto/model/settingsDTO.go
@@ -1,7 +1,7 @@
 package model
 
 type SettingsDTO struct {
-	PlannedCalories int    `json:"PlannedCalories"`
-	Gender          string `json:"Gender"`
-	WeightInKg      int    `json:"WeightInKg"`
+	PlannedCalories int     `json:"PlannedCalories"`
+	Gender          string  `json:"Gender"`
+	WeightInKg      float64 `json:"WeightInKg"`
 }
