+++ backend/tracker/infrastructure/entity/settings.go
@@ -1,10 +1,10 @@
 package entity
 
 type Settings struct {
-	PlannedCalories int    `json:"PlannedCalories"`
-	ProteinsInGrams int    `json:"ProteinsInGrams"`
-	FatsInGrams     int    `json:"FatsInGrams"`
-	CarbsInGrams    int    `json:"CarbsInGrams"`
-	Gender          string `json:"Gender" gorm:"type:enum('männlich', 'weiblich', 'divers')"`
-	WeightInKg      int    `json:"WeightInKg"`
+	PlannedCalories int     `json:"PlannedCalories"`
+	ProteinsInGrams int     `json:"ProteinsInGrams"`
+	FatsInGrams     int     `json:"FatsInGrams"`
+	CarbsInGrams    int     `json:"CarbsInGrams"`
+	Gender          string  `json:"Gender" gorm:"type:enum('männlich', 'weiblich', 'divers')"`
+	WeightInKg      float64 `json:"WeightInKg"`
 }
